Fix ErrorHandlingFilter config (#51)
In a downstream project generated from play-scala-template I created an
additional filter (an AccessLoggingFilter) and added it within
BaseFilters.

Having done so I found that the AccessLoggingFilter wasn't actually
instantiated.

I then checked whether ErrorHandlingFilter was even being instantiated.
It wasn't. Error handling was actually being done by the built in Play
error handler.

Although BaseFilters and BaseFiltersSpec themselves were correct in a sense,
the issue was that BaseFilters wasn't actually being wired in at runtime.

As per the docs at
https://www.playframework.com/documentation/2.6.x/Filters#Default-Filters,
when using runtime DI in Play 2.6, filters should be enabled via config.
For BaseFilters to have been used it would have to have just been called
"Filters" and been placed in the root package. From the above docs:
"Otherwise, if you have a Filters class in the root or have
play.http.filters defined explicitly, it will take precedence over the
EnabledFilters functionality described below."

But that would be a bit _too_ magic. Hence the removal of BaseFilters and the
addition of filters.enabled += to the config.

With this change made in the downstream project, both the AccessLoggingFilter
and ErrorHandlingFilter are instantiated as they should be.